Move PartitionDirect tablet into a separate library
