+++ ydb/core/grpc_services/ya.make
@@ -259,7 +259,6 @@ IF (OS_LINUX AND YDB_EMBEDDED_NBS_ENABLED)
     )
     PEERDIR(
         ydb/core/nbs/cloud/blockstore/libs/service
-        ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct
         ydb/core/nbs/cloud/blockstore/libs/storage/ss_proxy
         ydb/core/nbs/cloud/blockstore/public/api/protos
         ydb/core/nbs/cloud/storage/core/libs/common
