+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/ya.make
@@ -6,25 +6,11 @@ GENERATE_ENUM_SERIALIZATION(partition_direct_service.h)
 
 SRCS(
     ddisk_data_copier.cpp
-    delete_partition.cpp
     direct_block_group_impl.cpp
     direct_block_group.cpp
     erase_request.cpp
     fast_path_service.cpp
     flush_request.cpp
-    load_actor_adapter.cpp
-    part_add_host_to_dbg.cpp
-    part_database.cpp
-    part_initschema.cpp
-    part_loadstate.cpp
-    part_storepartitionids.cpp
-    part_storevolumeconfig.cpp
-    part_updatevchunkconfig.cpp
-    part_updatedirtymapstate.cpp
-    part_monitoring.cpp
-    partition_cleanup_actor.cpp
-    partition_direct_actor.cpp
-    partition_direct.cpp
     read_request_executor.cpp
     read_request_multiple_location.cpp
     read_request_single_location.cpp
@@ -69,6 +55,5 @@ RECURSE(
 
 RECURSE_FOR_TESTS(
     benchmark
-    partition_ut
     ut
 )
