+++ ydb/core/nbs/cloud/blockstore/libs/storage/volume/ya.make
@@ -10,7 +10,6 @@ PEERDIR(
     ydb/core/nbs/cloud/blockstore/libs/service
     ydb/core/nbs/cloud/blockstore/libs/storage/api
     ydb/core/nbs/cloud/blockstore/libs/storage/core
-    ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct
     ydb/core/nbs/cloud/storage/core/libs/common
 
     ydb/library/actors/core
