+++ ydb/core/nbs/cloud/blockstore/libs/storage/ya.make
@@ -2,6 +2,7 @@ RECURSE(
     api
     dbs_controller
     partition_direct
+    partition_direct_tablet
     storage_transport
     testlib
 )
